+++ app/Providers/DataTableServiceProvider.php
@@ -12,7 +12,6 @@ use App\Services\DataTable\Services\FilterService;
 use App\Services\DataTable\Services\SearchService;
 use App\Services\DataTable\Services\SessionService;
 use App\Services\DataTable\Services\SortService;
-use App\Services\DataTable\Services\StatsService;
 use App\Services\DataTable\Ui\DataTableComponentRegistry;
 use App\Services\DataTable\Ui\DataTableFilterComponentRegistry;
 use Illuminate\Support\ServiceProvider;
@@ -35,7 +34,6 @@ class DataTableServiceProvider extends ServiceProvider
         $this->app->singleton(SearchService::class);
         $this->app->singleton(FilterService::class);
         $this->app->singleton(SortService::class);
-        $this->app->singleton(StatsService::class);
         $this->app->singleton(SessionService::class);
         $this->app->singleton(DataTableComponentRegistry::class);
         $this->app->singleton(DataTableFilterComponentRegistry::class);