+++ resources/views/components/table/actions.blade.php
@@ -3,35 +3,49 @@
     'itemUuid' => '',
 ])
 
-<div class="flex items-center gap-2" wire:click.stop>
-    @foreach ($actions as $action)
-        @if (($action['key'] ?? null) === 'delete')
-            <x-ui.button
-                style="ghost"
-                color="error"
-                size="sm"
-                @click="$dispatch('confirm-modal', {
-                    title: '{{ addslashes($action['label'] ?? __('ui.actions.delete')) }}',
-                    message: '{{ addslashes(__('ui.modals.confirm.message')) }}',
-                    confirmAction: () => $wire.handleRowAction('{{ $action['key'] }}', '{{ $itemUuid }}')
-                })"
-            >
-                @if (isset($action['icon']))
-                    <x-ui.icon name="{{ $action['icon'] }}" class="h-4 w-4"></x-ui.icon>
-                @endif
+@php
+    use App\Constants\DataTable\DataTableUi;
+@endphp
+
+@if (count($actions) > 0)
+    <x-ui.dropdown placement="end" menu>
+        <x-slot:trigger>
+            <x-ui.button variant="ghost" size="sm" wire:click.stop>
+                <x-ui.icon name="{{ DataTableUi::ICON_THREE_DOTS }}" class="h-4 w-4"></x-ui.icon>
             </x-ui.button>
-        @else
-            <x-ui.button
-                style="{{ $action['variant'] ?? 'ghost' }}"
-                color="{{ $action['color'] ?? null }}"
-                size="sm"
-                wire:click.stop="handleRowAction('{{ $action['key'] }}', '{{ $itemUuid }}')"
-            >
-                @if (isset($action['icon']))
-                    <x-ui.icon name="{{ $action['icon'] }}" class="h-4 w-4"></x-ui.icon>
+        </x-slot:trigger>
+
+        @foreach ($actions as $action)
+            <li>
+                @if (($action['key'] ?? null) === 'delete')
+                    <button
+                        type="button"
+                        class="w-full text-left text-error"
+                        @click="$dispatch('confirm-modal', {
+                            title: '{{ addslashes($action['label'] ?? __('ui.actions.delete')) }}',
+                            message: '{{ addslashes(__('ui.modals.confirm.message')) }}',
+                            confirmAction: () => $wire.handleRowAction('{{ $action['key'] }}', '{{ $itemUuid }}')
+                        })"
+                    >
+                        @if (isset($action['icon']))
+                            <x-ui.icon name="{{ $action['icon'] }}" class="h-4 w-4 mr-2"></x-ui.icon>
+                        @endif
+                        {{ $action['label'] ?? __('ui.actions.delete') }}
+                    </button>
+                @else
+                    <button
+                        type="button"
+                        class="w-full text-left"
+                        wire:click.stop="handleRowAction('{{ $action['key'] }}', '{{ $itemUuid }}')"
+                    >
+                        @if (isset($action['icon']))
+                            <x-ui.icon name="{{ $action['icon'] }}" class="h-4 w-4 mr-2"></x-ui.icon>
+                        @endif
+                        {{ $action['label'] ?? '' }}
+                    </button>
                 @endif
-            </x-ui.button>
-        @endif
-    @endforeach
-</div>
+            </li>
+        @endforeach
+    </x-ui.dropdown>
+@endif